Moved httpconv package to typedvalues
